feat: staticify before comment

+++ ansible/files/scripts/static-ify.sh
@@ -3,8 +3,12 @@
 export DZL_PATH="/dzl/src/dzl/cli"
 cd "$DZL_PATH" || exit 1
 
-nohup node ./bin/dzl.js serve &
-SERVER_PID=$!
+if nc -z 127.0.0.1 8088 ; then
+  echo "Server is already up, skipping..."
+else
+  nohup node ./bin/dzl.js serve --config=/dzl/conf/agent-official.config.js &
+  SERVER_PID=$!
+fi
 
 rm -rf dist/
 mkdir dist/
@@ -28,11 +32,13 @@ curl http://localhost:8088/by-url.js > dist/by-url.js
 curl http://localhost:8088/comparison.js > dist/comparison.js
 curl http://localhost:8088/dashboard.js > dist/dashboard.js
 
-kill $SERVER_PID
+if [[ -n $SERVER_PID ]]; then
+  kill $SERVER_PID
+fi
 
 echo "lh-dzl-${PR_ID:-master}.surge.sh" > dist/CNAME
 
 if which surge; then
-  cd dist && surge .
+  surge ./dist
 fi
